+++ packages/hedera-wallet-snap/packages/snap/src/facades/hts/CreateTokenFacade.ts
@@ -21,12 +21,12 @@
 import { providerErrors } from '@metamask/rpc-errors';
 import { divider, heading, text } from '@metamask/snaps-ui';
 import _ from 'lodash';
-import { HederaClientImplFactory } from '../client/HederaClientImplFactory';
-import { CreateTokenCommand } from '../commands/hts/CreateTokenCommand';
-import { TxReceipt } from '../types/hedera';
-import { CreateTokenRequestParams } from '../types/params';
-import { SnapDialogParams, WalletSnapParams } from '../types/state';
-import { SnapUtils } from '../utils/SnapUtils';
+import { HederaClientImplFactory } from '../../client/HederaClientImplFactory';
+import { CreateTokenCommand } from '../../commands/hts/CreateTokenCommand';
+import { TxReceipt } from '../../types/hedera';
+import { CreateTokenRequestParams } from '../../types/params';
+import { SnapDialogParams, WalletSnapParams } from '../../types/state';
+import { SnapUtils } from '../../utils/SnapUtils';
 
 export class CreateTokenFacade {
   /**
