+++ packages/hedera-wallet-snap/packages/snap/src/facades/hts/MintTokenFacade.ts
@@ -21,25 +21,27 @@
 import { providerErrors } from '@metamask/rpc-errors';
 import { divider, heading, text } from '@metamask/snaps-ui';
 import _ from 'lodash';
-import { HederaClientImplFactory } from '../client/HederaClientImplFactory';
-import { BurnTokenCommand } from '../commands/hts/BurnTokenCommand';
-import { TxReceipt } from '../types/hedera';
-import { BurnTokenRequestParams } from '../types/params';
-import { SnapDialogParams, WalletSnapParams } from '../types/state';
-import { CryptoUtils } from '../utils/CryptoUtils';
-import { SnapUtils } from '../utils/SnapUtils';
+import { HederaClientImplFactory } from '../../client/HederaClientImplFactory';
+import { MintTokenCommand } from '../../commands/hts/MintTokenCommand';
+import { TxReceipt } from '../../types/hedera';
+import { MintTokenRequestParams } from '../../types/params';
+import { SnapDialogParams, WalletSnapParams } from '../../types/state';
+import { CryptoUtils } from '../../utils/CryptoUtils';
+import { SnapUtils } from '../../utils/SnapUtils';
 
-export class BurnTokenFacade {
+export class MintTokenFacade {
   /**
-   * Burns fungible and non-fungible tokens owned by the Treasury Account.
+   * Minting fungible token allows you to increase the total supply of the token. Minting a
+   * non-fungible token creates an NFT with its unique metadata for the class of NFTs defined by
+   * the token ID. The Supply Key must sign the transaction.
    *
    * @param walletSnapParams - Wallet snap params.
-   * @param burnTokenRequestParams - Parameters for burning a token.
+   * @param mintTokenRequestParams - Parameters for minting a token.
    * @returns Receipt of the transaction.
    */
-  public static async burnToken(
+  public static async mintToken(
     walletSnapParams: WalletSnapParams,
-    burnTokenRequestParams: BurnTokenRequestParams,
+    mintTokenRequestParams: MintTokenRequestParams,
   ): Promise<TxReceipt> {
     const { origin, state } = walletSnapParams;
 
@@ -50,8 +52,8 @@ export class BurnTokenFacade {
       assetType,
       tokenId,
       amount,
-      serialNumbers = [],
-    } = burnTokenRequestParams;
+      metadata = [],
+    } = mintTokenRequestParams;
 
     const { privateKey, curve } =
       state.accountState[hederaEvmAddress][network].keyStore;
@@ -59,12 +61,12 @@ export class BurnTokenFacade {
     let txReceipt = {} as TxReceipt;
     try {
       const panelToShow = [
-        heading('Burn token'),
+        heading('Mint token'),
         text(
-          'Learn more about burning tokens [here](https://docs.hedera.com/hedera/sdks-and-apis/sdks/readme-1/burn-a-token)',
+          'Learn more about minting tokens [here](https://docs.hedera.com/hedera/sdks-and-apis/sdks/readme-1/mint-a-token)',
         ),
         text(
-          `You are about to burn a ${
+          `You are about to mint a ${
             assetType === 'TOKEN' ? 'Fungible Token' : 'Non Fungible Token(NFT)'
           } with the following details:`,
         ),
@@ -73,16 +75,16 @@ export class BurnTokenFacade {
       ];
       if (assetType === 'NFT') {
         panelToShow.push(
-          text(`Amount to burn: ${serialNumbers.length}`),
-          text(`Serial Numbers: `),
+          text(`Amount to mint: ${metadata.length}`),
+          text(`Metadata: `),
         );
         panelToShow.push(divider());
-        for (const serialNumber of serialNumbers) {
-          panelToShow.push(text(`🔥 ${serialNumber}`));
+        for (const data of metadata) {
+          panelToShow.push(text(`🚀 ${data}`));
         }
         panelToShow.push(divider());
       } else {
-        panelToShow.push(text(`Amount to burn: ${amount as number}`));
+        panelToShow.push(text(`Amount to mint: ${amount as number}`));
       }
       const tokenInfo = await CryptoUtils.getTokenById(tokenId, mirrorNodeUrl);
       if (_.isEmpty(tokenInfo)) {
@@ -91,7 +93,7 @@ export class BurnTokenFacade {
         throw new Error(errMessage);
       }
       panelToShow.push(
-        text(`Burn from Treasury account: ${tokenInfo.treasury_account_id}`),
+        text(`Mint to Treasury account: ${tokenInfo.treasury_account_id}`),
       );
       panelToShow.push(text(`Asset Name: ${tokenInfo.name}`));
       panelToShow.push(text(`Asset Type: ${tokenInfo.type}`));
@@ -134,10 +136,10 @@ export class BurnTokenFacade {
       if (hederaClient === null) {
         throw new Error('hedera client returned null');
       }
-      const mintTokenCommand = new BurnTokenCommand(
+      const mintTokenCommand = new MintTokenCommand(
         assetType,
         tokenId,
-        serialNumbers,
+        metadata,
         assetType === 'TOKEN'
           ? Number(amount) * Math.pow(10, Number(tokenInfo.decimals))
           : amount,
@@ -149,7 +151,7 @@ export class BurnTokenFacade {
       }
       txReceipt = await mintTokenCommand.execute(hederaClient.getClient());
     } catch (error: any) {
-      const errMessage = `Error while trying to burn tokens: ${String(error)}`;
+      const errMessage = `Error while trying to mint tokens: ${String(error)}`;
       console.error(errMessage);
       throw providerErrors.unsupportedMethod(errMessage);
     }
